Backend: fixed nasty issue with int_t being int rather than long long

+++ include/isaac/types.h
@@ -8,7 +8,7 @@
 namespace isaac
 {
 
-typedef int int_t;
+typedef long long int_t;
 
 struct ISAACAPI size4
 {
